+++ src/ttsim/main_target.py
@@ -92,6 +92,9 @@ class FailIf(MainTargetABC):
 class Results(MainTargetABC):
     df_with_mapper: str = "results__df_with_mapper"
     df_with_nested_columns: str = "results__df_with_nested_columns"
+    df_with_qname_columns: str = "results__df_with_qname_columns"
+    flat: str = "results__flat"
+    qname: str = "results__qname"
     tree: str = "results__tree"
 
 
@@ -157,7 +160,9 @@ class DfAndMapper(MainTargetABC):
 class InputData(MainTargetABC):
     df_and_mapper: type[DfAndMapper] = field(default=DfAndMapper)
     df_with_nested_columns: str = "input_data__df_with_nested_columns"
+    df_with_qname_columns: str = "input_data__df_with_qname_columns"
     flat: str = "input_data__flat"
+    qname: str = "input_data__qname"
     sort_indices: str = "input_data__sort_indices"
     tree: str = "input_data__tree"